refactor: 기능별 함수를 모아둔 파일 분리 및 불러오기 변경
